+++ src/main/java/org/eclipse/keyple/core/service/CardSelectionManagerAdapter.java
@@ -11,6 +11,11 @@
  ************************************************************************************** */
 package org.eclipse.keyple.core.service;
 
+import static org.eclipse.keyple.core.service.InternalDto.*;
+
+import com.google.gson.JsonArray;
+import com.google.gson.JsonObject;
+import com.google.gson.reflect.TypeToken;
 import java.util.ArrayList;
 import java.util.List;
 import org.calypsonet.terminal.card.*;
@@ -28,6 +33,9 @@ import org.calypsonet.terminal.reader.selection.ScheduledCardSelectionsResponse;
 import org.calypsonet.terminal.reader.selection.spi.CardSelection;
 import org.calypsonet.terminal.reader.selection.spi.SmartCard;
 import org.eclipse.keyple.core.util.Assert;
+import org.eclipse.keyple.core.util.json.JsonUtil;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * (package-private) <br>
@@ -37,10 +45,23 @@ import org.eclipse.keyple.core.util.Assert;
  */
 final class CardSelectionManagerAdapter implements CardSelectionManager {
 
+  private static final Logger logger = LoggerFactory.getLogger(CardSelectionManagerAdapter.class);
+  private static final String DETECTION_MODE = "detectionMode";
+  private static final String NOTIFICATION_MODE = "notificationMode";
+  private static final String MULTI_SELECTION_PROCESSING = "multiSelectionProcessing";
+  private static final String CHANNEL_CONTROL = "channelControl";
+  private static final String CARD_SELECTIONS_TYPES = "cardSelectionsTypes";
+  private static final String CARD_SELECTIONS = "cardSelections";
+  private static final String DEFAULT_CARD_SELECTIONS = "defaultCardSelections";
+
   private final List<CardSelectionSpi> cardSelections;
   private final List<CardSelectionRequestSpi> cardSelectionRequests;
   private MultiSelectionProcessing multiSelectionProcessing;
   private ChannelControl channelControl = ChannelControl.KEEP_OPEN;
+  private ObservableCardReader.DetectionMode detectionMode =
+      ObservableCardReader.DetectionMode.REPEATING;
+  private ObservableCardReader.NotificationMode notificationMode =
+      ObservableCardReader.NotificationMode.ALWAYS;
 
   /**
    * (package-private) <br>
@@ -95,6 +116,106 @@ final class CardSelectionManagerAdapter implements CardSelectionManager {
   /**
    * {@inheritDoc}
    *
+   * @since 2.1.1
+   */
+  @Override
+  public String exportCardSelectionScenario(
+      ObservableCardReader.DetectionMode detectionMode,
+      ObservableCardReader.NotificationMode notificationMode) {
+
+    Assert.getInstance()
+        .notNull(detectionMode, DETECTION_MODE)
+        .notNull(notificationMode, NOTIFICATION_MODE);
+
+    JsonObject jsonObject = new JsonObject();
+
+    // Basic fields
+    jsonObject.addProperty(MULTI_SELECTION_PROCESSING, multiSelectionProcessing.name());
+    jsonObject.addProperty(CHANNEL_CONTROL, channelControl.name());
+    jsonObject.addProperty(DETECTION_MODE, detectionMode.name());
+    jsonObject.addProperty(NOTIFICATION_MODE, notificationMode.name());
+
+    // Original card selections
+    List<String> cardSelectionsTypes = new ArrayList<String>(cardSelections.size());
+    for (CardSelectionSpi cardSelection : cardSelections) {
+      cardSelectionsTypes.add(cardSelection.getClass().getName());
+    }
+    jsonObject.add(CARD_SELECTIONS_TYPES, JsonUtil.getParser().toJsonTree(cardSelectionsTypes));
+    jsonObject.add(CARD_SELECTIONS, JsonUtil.getParser().toJsonTree(cardSelections));
+
+    // Default card selections
+    List<CardSelectionAdapter> defaultCardSelections =
+        new ArrayList<CardSelectionAdapter>(cardSelections.size());
+    for (CardSelectionSpi cardSelection : cardSelections) {
+      defaultCardSelections.add(new CardSelectionAdapter(cardSelection));
+    }
+    jsonObject.add(DEFAULT_CARD_SELECTIONS, JsonUtil.getParser().toJsonTree(defaultCardSelections));
+
+    return jsonObject.toString();
+  }
+
+  /**
+   * {@inheritDoc}
+   *
+   * @since 2.1.1
+   */
+  @Override
+  public int importCardSelectionScenario(String cardSelectionScenario) {
+
+    JsonObject jsonObject = JsonUtil.getParser().fromJson(cardSelectionScenario, JsonObject.class);
+
+    // Basic fields
+    multiSelectionProcessing =
+        MultiSelectionProcessing.valueOf(jsonObject.get(MULTI_SELECTION_PROCESSING).getAsString());
+    channelControl = ChannelControl.valueOf(jsonObject.get(CHANNEL_CONTROL).getAsString());
+    detectionMode =
+        ObservableCardReader.DetectionMode.valueOf(jsonObject.get(DETECTION_MODE).getAsString());
+    notificationMode =
+        ObservableCardReader.NotificationMode.valueOf(
+            jsonObject.get(NOTIFICATION_MODE).getAsString());
+
+    // Card selections
+    List<String> cardSelectionsTypes =
+        JsonUtil.getParser()
+            .fromJson(
+                jsonObject.get(CARD_SELECTIONS_TYPES).getAsJsonArray(),
+                new TypeToken<ArrayList<String>>() {}.getType());
+    JsonArray cardSelectionsJsonArray = jsonObject.get(CARD_SELECTIONS).getAsJsonArray();
+    JsonArray defaultCardSelectionsJsonArray =
+        jsonObject.get(DEFAULT_CARD_SELECTIONS).getAsJsonArray();
+
+    // Clear the current list of card selections
+    cardSelections.clear();
+    cardSelectionRequests.clear();
+
+    int index = 0;
+    for (int i = 0; i < cardSelectionsTypes.size(); i++) {
+      CardSelection cardSelection;
+      try {
+        Class<?> classOfCardSelection = Class.forName(cardSelectionsTypes.get(i));
+        // Original card selection
+        cardSelection =
+            (CardSelection)
+                JsonUtil.getParser().fromJson(cardSelectionsJsonArray.get(i), classOfCardSelection);
+      } catch (ClassNotFoundException e) {
+        // Default card selection
+        logger.warn(
+            "Original CardSelection type '{}' not found. Use default type '{}' for deserialization.",
+            cardSelectionsTypes.get(i),
+            CardSelectionAdapter.class.getName());
+        cardSelection =
+            JsonUtil.getParser()
+                .fromJson(defaultCardSelectionsJsonArray.get(i), CardSelectionAdapter.class);
+      }
+      // Prepare selection
+      index = prepareSelection(cardSelection);
+    }
+    return index;
+  }
+
+  /**
+   * {@inheritDoc}
+   *
    * @since 2.0.0
    */
   @Override
@@ -133,8 +254,8 @@ final class CardSelectionManagerAdapter implements CardSelectionManager {
 
     Assert.getInstance()
         .notNull(observableCardReader, "observableCardReader")
-        .notNull(detectionMode, "detectionMode")
-        .notNull(notificationMode, "notificationMode");
+        .notNull(detectionMode, DETECTION_MODE)
+        .notNull(notificationMode, NOTIFICATION_MODE);
 
     CardSelectionScenarioAdapter cardSelectionScenario =
         new CardSelectionScenarioAdapter(
@@ -148,6 +269,16 @@ final class CardSelectionManagerAdapter implements CardSelectionManager {
     } else {
       throw new IllegalArgumentException("Not a Keyple reader implementation.");
     }
+  }
+
+  /**
+   * {@inheritDoc}
+   *
+   * @since 2.1.1
+   */
+  @Override
+  public void scheduleCardSelectionScenario(ObservableCardReader observableCardReader) {
+    scheduleCardSelectionScenario(observableCardReader, detectionMode, notificationMode);
   }
 
   /**
@@ -189,7 +320,7 @@ final class CardSelectionManagerAdapter implements CardSelectionManager {
     for (CardSelectionResponseApi cardSelectionResponse : cardSelectionResponses) {
       if (cardSelectionResponse.hasMatched()) {
         // invoke the parse method defined by the card extension to retrieve the smart card
-        SmartCard smartCard = null;
+        SmartCard smartCard;
         try {
           smartCard = (SmartCard) cardSelections.get(index).parse(cardSelectionResponse);
         } catch (ParseException e) {