+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- The possibility to import/export a card selection scenario.
 ### Fixed
 - Logging format for distributed service.
+### Upgraded
+- "Calypsonet Terminal Reader API" to version `1.1.0`.
+- "Keyple Util Library" to version `2.2.0`.
 
 ## [2.1.0] - 2022-07-25
 ### Added